feat(optimiser): implement OptimiserButton, OptimiserForm, and OptimiserHeader components with styling for enhanced user interaction
