+++ website/src/views/optimiser/OptimiserForm.scss
@@ -1,0 +1,301 @@
+@import '~styles/utils/modules-entry.scss';
+
+.lessonTag {
+  display: inline-flex;
+  align-items: center;
+  padding: 0.25rem 0.5rem;
+  border-bottom: 3px solid;
+  border-radius: 0.25rem;
+  font-weight: 500;
+  font-size: 0.8rem;
+  transition: transform 0.15s ease, filter 0.15s ease;
+  gap: 0.25rem;
+
+  &:hover {
+    filter: brightness(0.8);
+  }
+} 
+
+.tag {
+    cursor: pointer;
+}
+
+// Main content section
+.mainContent {
+  display: flex;
+  flex-direction: column;
+  margin-top: 2rem;
+}
+
+// Section headers
+.sectionHeader {
+  display: flex;
+  align-items: center;
+  margin-bottom: 1rem;
+  font-size: 1rem;
+  gap: 0.5rem;
+}
+
+// Lesson selection section
+.lessonButtons {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.lessonButton {
+  display: inline-flex;
+  align-items: center;
+  padding: 0.5rem 0.5rem;
+  border-top: none;
+  border-right: none;
+  border-left: none;
+  border-radius: 0.25rem;
+  font-weight: 500;
+  font-size: 0.8rem;
+  cursor: pointer;
+  transition: all 0.15s ease;
+  gap: 0.25rem;
+
+  &.selected {
+    opacity: 1;
+    filter: brightness(1);
+    transform: scale(1);
+  }
+
+  &.unselected {
+    opacity: 0.6;
+    filter: brightness(0.8);
+    transform: scale(0.98);
+  }
+
+  &:hover {
+    opacity: 1 !important;
+    filter: brightness(1.1) !important;
+    transform: scale(1.02) !important;
+  }
+}
+
+.lessonButtonText {
+  font-weight: bold;
+}
+
+// Free days section
+.freeDaysSection {
+  display: flex;
+  align-items: center;
+  margin-top: 2rem;
+  font-size: 1rem;
+  gap: 0.5rem;
+}
+
+.freeDaysButtons {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  margin-top: 1rem;
+  gap: 0.5rem;
+}
+
+// Conflict warning
+.conflictWarning {
+  display: flex;
+  flex-direction: column;
+  padding: 1rem;
+  margin-top: 1rem;
+  border: 1px solid rgba(255, 81, 56, 0.3);
+  border-radius: 0.5rem;
+  background-color: rgba(255, 81, 56, 0.1);
+  gap: 0.5rem;
+}
+
+.conflictHeader {
+  display: flex;
+  align-items: center;
+  font-weight: bold;
+  font-size: 1rem;
+  color: #ff5138;
+  gap: 0.5rem;
+}
+
+.conflictDescription {
+  font-size: 0.9rem;
+  color: #69707a;
+}
+
+.conflictItem {
+  margin-left: 1rem;
+  font-weight: 500;
+  font-size: 0.9rem;
+  color: #ff5138;
+}
+
+.conflictFooter {
+  margin-top: 0.5rem;
+  font-style: italic;
+  font-size: 0.8rem;
+  color: #69707a;
+}
+
+// Time controls section
+.timeControls {
+  display: flex;
+  flex-wrap: wrap;
+  margin-top: 2rem;
+  gap: 3rem;
+
+  @include media-breakpoint-down(md) {
+    flex-direction: column;
+    gap: 1.5rem;
+  }
+}
+
+.timeControlWrapper {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.timeControlGroup {
+  display: flex;
+  flex-direction: column;
+  width: auto;
+  min-width: 200px;
+  gap: 0.5rem;
+
+  @include media-breakpoint-down(md) {
+    width: 100%;
+    min-width: auto;
+  }
+}
+
+.timeControlHeader {
+  display: flex;
+  align-items: center;
+  font-size: 1rem;
+  gap: 0.5rem;
+}
+
+.timeControlRow {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.timeSelect {
+  width: 5rem;
+  padding: 0.25rem 0.75rem;
+  border: 1px solid var(--gray-lighter);
+  border-radius: 0.25rem;
+  outline: none;
+  font-size: 1.2rem;
+  font-family: monospace;
+  color: inherit;
+  background-color: transparent;
+}
+
+.timeLabel {
+  font-weight: 500;
+  font-size: 1.2rem;
+  font-family: monospace;
+  color: inherit;
+}
+
+// Lunch controls section
+.lunchControls {
+  display: flex;
+  flex-wrap: wrap;
+  margin-top: 2rem;
+  gap: 3rem;
+
+  @include media-breakpoint-down(md) {
+    flex-direction: column;
+    gap: 1.5rem;
+  }
+}
+
+.lunchControlGroup {
+  display: flex;
+  flex-direction: column;
+  width: auto;
+  min-width: 200px;
+  gap: 0.5rem;
+
+  @include media-breakpoint-down(md) {
+    width: 100%;
+    min-width: auto;
+  }
+}
+
+.lunchControlHeader {
+  display: flex;
+  align-items: center;
+  font-size: 1rem;
+  gap: 0.5rem;
+}
+
+.lunchControlRow {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.lunchSelect {
+  width: 5rem;
+  padding: 0.25rem 0.75rem;
+  border: 1px solid var(--gray-lighter);
+  border-radius: 0.25rem;
+  outline: none;
+  font-size: 1.2rem;
+  font-family: monospace;
+  color: inherit;
+  background-color: transparent;
+}
+
+.lunchTimeLabel {
+  font-weight: 500;
+  font-size: 1.2rem;
+  font-family: monospace;
+  color: inherit;
+}
+
+.lunchTimeSeparator {
+  margin: 0 1.5rem;
+  font-size: 1rem;
+  color: #69707a;
+}
+
+.noLessonsFound {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  padding: 1.5rem;
+  margin-top: 0.5rem;
+  border: 1px solid rgba(255, 193, 7, 0.3);
+  border-radius: 0.75rem;
+  background-color: rgba(255, 193, 7, 0.1);
+  gap: 1rem;
+}
+
+.noLessonsHeader {
+  display: flex;
+  align-items: center;
+  font-weight: bold;
+  font-size: 1.1rem;
+  color: #ff8c00;
+  gap: 0.75rem;
+}
+
+.noLessonsDescription {
+  font-size: 0.95rem;
+  line-height: 1.5;
+  color: rgba(255, 193, 7, 0.5);
+}
+
+// Icon styling
+.infoIcon {
+  color: #69707a;
+}
